--- app/RootStyleRegistry.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { useStyledComponentsRegistry } from '../lib/styled-components';
-import { useServerInsertedHTML } from 'next/navigation';
-
-export default function RootStyleRegistry({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [StyledComponentsRegistry, styledComponentsFlushEffect] =
-    useStyledComponentsRegistry();
-
-  useServerInsertedHTML(() => {
-    return <>{styledComponentsFlushEffect()}</>;
-  });
-
-  return <StyledComponentsRegistry>{children}</StyledComponentsRegistry>;
-}
