+++ app/head.tsx
@@ -2,12 +2,12 @@ export default function Head() {
   return (
     <>
       <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <link
+      {/* <link
         rel="stylesheet"
         as="style"
         crossOrigin="anonymous"
         href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/static/pretendard-dynamic-subset.css"
-      />
+      /> */}
       <meta
         name="description"
         content="A playground to explore new Next.js 13 app directory features such as nested layouts, instant loading states, streaming, and component level data fetching."
