+++ app/page.tsx
@@ -1,48 +1,4 @@
-'use client';
-
-import styled from 'styled-components';
-
-const Container = styled.div`
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  gap: 1.5rem /* 24px */;
-`;
-
-const SkeletonInner = styled.div`
-  padding: 1rem /* 16px */;
-  background-color: rgb(24 24 27 / 0.8);
-  border-radius: 1rem /* 16px */;
-`;
-
-const SkeletonImg = styled.div`
-  height: 3.5rem /* 56px */;
-  border-radius: 0.5rem /* 8px */;
-  background-color: rgb(63 63 70 / 1);
-`;
-
-const SkeletonBtn = styled.div`
-  margin-top: 0.75rem /* 12px */;
-  width: 25%;
-  height: 0.75rem /* 12px */;
-  border-radius: 0.5rem /* 8px */;
-  background-color: rgb(255 0 128 / 1);
-`;
-
-const SkeletonLineOne = styled.div`
-  margin-top: 0.75rem /* 12px */;
-  height: 0.75rem /* 12px */;
-  width: 91.666667%;
-  border-radius: 0.5rem /* 8px */;
-  background-color: rgb(63 63 70 / 1);
-`;
-
-const SkeletonLineTwo = styled.div`
-  margin-top: 0.75rem /* 12px */;
-  height: 0.75rem /* 12px */;
-  width: 66.666667%;
-  border-radius: 0.5rem /* 8px */;
-  background-color: rgb(63 63 70 / 1);
-`;
+import { Container, SkeletonInner, SkeletonImg, SkeletonBtn, SkeletonLineOne, SkeletonLineTwo } from "./page.styles";
 
 const Skeleton = () => (
   <SkeletonInner>
