+++ package.json
@@ -12,13 +12,13 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "@tanstack/react-query": "^4.16.1",
+    "@tanstack/react-query": "^4.20.4",
     "axios": "^1.1.3",
-    "next": "13.0.5",
+    "next": "13.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "styled-components": "6.0.0-beta.6",
-    "typescript": "4.9.3"
+    "typescript": "4.9.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.2",
@@ -40,7 +40,7 @@
     "babel-loader": "^8.3.0",
     "chromatic": "^6.11.4",
     "eslint": "8.28.0",
-    "eslint-config-next": "13.0.4",
+    "eslint-config-next": "13.1.1",
     "msw": "^0.49.0"
   },
   "packageManager": "yarn@3.3.0"
