--- pages/_app.tsx
@@ -1,18 +1,0 @@
-import { AppProps } from 'next/app';
-import Head from 'next/head';
-import '../styles/globals.css';
-
-function CustomApp({ Component, pageProps }: AppProps) {
-  return (
-    <>
-      <Head>
-        <title>Welcome to universe/dashboard!</title>
-      </Head>
-      <main className="app">
-        <Component {...pageProps} />
-      </main>
-    </>
-  );
-}
-
-export default CustomApp;
